requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
